Move more mocks into __mocks__
