+++ __mocks__/@react-native-community/geolocation.ts
@@ -1,0 +1,8 @@
+export default ( {
+  getCurrentPosition: jest.fn( ),
+  watchPosition: jest.fn( ( ) => 0 ),
+  clearWatch: jest.fn( ),
+  setRNConfiguration: jest.fn( ( ) => ( {
+    skipPermissionRequests: true
+  } ) )
+} );
