+++ __mocks__/react-native-config.ts
@@ -1,0 +1,6 @@
+export default ( {
+  OAUTH_CLIENT_ID: process.env.OAUTH_CLIENT_ID,
+  OAUTH_CLIENT_SECRET: process.env.OAUTH_CLIENT_SECRET,
+  JWT_ANONYMOUS_API_SECRET: process.env.JWT_ANONYMOUS_API_SECRET,
+  API_URL: process.env.API_URL
+} );
