+++ __mocks__/react-native-exif-reader.ts
@@ -1,0 +1,2 @@
+export const readExif = jest.fn( );
+export const writeLocation = jest.fn( );
