+++ __mocks__/react-native-geocoder-reborn.ts
@@ -1,0 +1,8 @@
+export default ( {
+  geocodePosition: jest.fn( coord => [
+    `Somewhere near ${coord.lat}, ${coord.lng}`,
+    "Somewhere",
+    "Somewheria",
+    "SW"
+  ] )
+} );
