+++ __mocks__/react-native-orientation-locker.ts
@@ -1,0 +1,11 @@
+export default ( {
+  addDeviceOrientationListener: jest.fn( ),
+  addEventListener: jest.fn( ),
+  getDeviceOrientation: jest.fn( ),
+  getInitialOrientation: jest.fn( ),
+  getOrientation: jest.fn( ),
+  lockToPortrait: jest.fn( ),
+  removeEventListener: jest.fn( ),
+  removeOrientationListener: jest.fn( ),
+  unlockAllOrientations: jest.fn( )
+} );
